Remove operating point subtractions from TS model
The operating points used to be subtracted from the input signal when
identifying the linear models and performing simulations. This is not
necessary, as model operating region is defined by the W (weight)
matrix, and linear model behaves identically regardless of it's
operating point.

Additionally, this subtraction and addition of the operating point
resulted in strange spikes in the final summed output. Now that the
model is identified and used with absolute input and output values, this
problem is gone.